revert writing static version to mix file
